refactor(api): move acceptance test to right file
